Split the velocity dispersion stuff into its own file
