+++ libgadget/Makefile
@@ -98,7 +98,7 @@ GADGET_OBJS =  \
 	 run.o drift.o stats.o \
 	 timestep.o init.o checkpoint.o \
 	 sfr_eff.o cooling.o cooling_rates.o cooling_uvfluc.o cooling_qso_lightup.o \
-	 winds.o density.o metal_return.o \
+	 winds.o veldisp.o density.o metal_return.o \
 	 treewalk.o cosmology.o \
 	 gravshort-tree.o gravshort-pair.o hydra.o  timefac.o \
 	 gravpm.o powerspectrum.o \
