+++ libgadget/blackhole.h
@@ -1,10 +1,8 @@
 #ifndef __BLACKHOLE_H
 #define __BLACKHOLE_H
 #include "utils/paramset.h"
-#include "timestep.h"
 #include "forcetree.h"
 #include "physconst.h"
-#include "density.h"
 
 enum BlackHoleFeedbackMethod {
      BH_FEEDBACK_TOPHAT   = 0x2,
@@ -26,9 +24,6 @@ void blackhole(const ActiveParticles * act, double atime, Cosmology * CP, ForceT
 
 /* Make a black hole from the particle at index. */
 void blackhole_make_one(int index, const double atime);
-
-/* Compute the DM velocity dispersion for black holes*/
-void blackhole_veldisp(const ActiveParticles * act, Cosmology * CP, ForceTree * tree, struct kick_factor_data * kf);
 
 /* Decide whether black hole repositioning is enabled. */
 int BHGetRepositionEnabled(void);
