+++ libgadget/winds.c
@@ -37,10 +37,6 @@ static struct WindParams
     double WindThermalFactor;
 } wind_params;
 
-#define NWINDHSML 5 /* Number of densities to evaluate for wind weight ngbiter*/
-#define NUMDMNGB 40 /*Number of DM ngb to evaluate vel dispersion */
-#define MAXDMDEVIATION 2
-
 typedef struct {
     TreeWalkQueryBase base;
     MyIDType ID;
@@ -588,299 +584,4 @@ winds_make_after_sf(int i, double sm, double vdisp, double atime)
         wind_do_kick(i, vel, utherm, atime);
     }
     return 0;
-}
-
-
-/* Code to compute velocity dispersions*/
-
-typedef struct {
-    TreeWalkQueryBase base;
-    double Mass;
-    double DMRadius[NWINDHSML];
-    double Vel[3];
-} TreeWalkQueryWindVDisp;
-
-typedef struct {
-    TreeWalkResultBase base;
-    double V1sum[NWINDHSML][3];
-    double V2sum[NWINDHSML];
-    double Ngb[NWINDHSML];
-    int alignment; /* Ensure alignment*/
-    int maxcmpte;
-} TreeWalkResultWindVDisp;
-
-struct WindVDispPriv {
-    double Time;
-    double hubble;
-    struct kick_factor_data kf;
-    double ddrift;
-    /* Lower and upper bounds on smoothing length*/
-    MyFloat *Left, *Right, *DMRadius;
-    /* Maximum index where NumNgb is valid. */
-    int * maxcmpte;
-    MyFloat (* V2sum)[NWINDHSML];
-    MyFloat (* V1sum)[NWINDHSML][3];
-    MyFloat (* Ngb) [NWINDHSML];
-};
-#define WINDV_GET_PRIV(tw) ((struct WindVDispPriv *) (tw->priv))
-
-static inline double
-vdispeffdmradius(int place, int i, TreeWalk * tw)
-{
-    double left = WINDV_GET_PRIV(tw)->Left[place];
-    double right = WINDV_GET_PRIV(tw)->Right[place];
-    /*The asymmetry is because it is free to compute extra densities for h < Hsml, but not for h > Hsml*/
-    if (right > 0.99*tw->tree->BoxSize){
-        right = WINDV_GET_PRIV(tw)->DMRadius[place];
-    }
-    if(left == 0)
-        left = 0.1 * WINDV_GET_PRIV(tw)->DMRadius[place];
-    /*Evenly split in volume*/
-    double rvol = pow(right, 3);
-    double lvol = pow(left, 3);
-    return pow((1.0*i+1)/(1.0*NWINDHSML+1) * (rvol - lvol) + lvol, 1./3);
-}
-
-static void
-wind_vdisp_copy(int place, TreeWalkQueryWindVDisp * input, TreeWalk * tw)
-{
-    input->Mass = P[place].Mass;
-    int i;
-    for(i=0; i<3; i++)
-        input->Vel[i] = P[place].Vel[i];
-    for(i = 0; i<NWINDHSML; i++){
-        input->DMRadius[i]=vdispeffdmradius(place,i,tw);
-    }
-}
-
-static void
-wind_vdisp_ngbiter(TreeWalkQueryWindVDisp * I,
-        TreeWalkResultWindVDisp * O,
-        TreeWalkNgbIterWind * iter,
-        LocalTreeWalk * lv)
-{
-    /* this evaluator walks the tree and sums the total mass of surrounding gas
-     * particles as described in VS08. */
-    /* it also calculates the velocity dispersion of the nearest 40 DM or gas particles */
-    if(iter->base.other == -1) {
-        iter->base.Hsml = I->DMRadius[NWINDHSML-1];
-        iter->base.mask = DMMASK; /* gas and dm */
-        iter->base.symmetric = NGB_TREEFIND_ASYMMETRIC;
-        O->maxcmpte = NWINDHSML;
-        return;
-    }
-
-    int other = iter->base.other;
-    double r = iter->base.r;
-    double * dist = iter->base.dist;
-
-    int i;
-    const double atime = WINDV_GET_PRIV(lv->tw)->Time;
-    for (i = 0; i < O->maxcmpte; i++) {
-        if(r < I->DMRadius[i]) {
-            O->Ngb[i] += 1;
-            int d;
-            MyFloat VelPred[3];
-            DM_VelPred(other, VelPred, &WINDV_GET_PRIV(lv->tw)->kf);
-            for(d = 0; d < 3; d ++) {
-                /* Add hubble flow to relative velocity. Use predicted velocity to current time.
-                 * The I particle is active so always at current time.*/
-                double vel = VelPred[d] - I->Vel[d] + WINDV_GET_PRIV(lv->tw)->hubble * atime * atime * dist[d];
-                O->V1sum[i][d] += vel;
-                O->V2sum[i] += vel * vel;
-            }
-        }
-    }
-
-    for(i = 0; i<NWINDHSML; i++){
-        if(O->Ngb[i] > NUMDMNGB){
-            O->maxcmpte = i+1;
-            iter->base.Hsml = I->DMRadius[i];
-            break;
-        }
-    }
-    /*
-    message(1, "ThisTask = %d %ld ngb=%d NGB=%d TotalWeight=%g V2sum=%g V1sum=%g %g %g\n",
-    ThisTask, I->ID, numngb, O->Ngb, O->TotalWeight, O->V2sum,
-    O->V1sum[0], O->V1sum[1], O->V1sum[2]);
-    */
-}
-
-static void
-wind_vdisp_reduce(int place, TreeWalkResultWindVDisp * O, enum TreeWalkReduceMode mode, TreeWalk * tw)
-{
-    int pi = P[place].PI;
-    int i;
-    if(mode == 0 || WINDV_GET_PRIV(tw)->maxcmpte[pi] > O->maxcmpte)
-        WINDV_GET_PRIV(tw)->maxcmpte[pi] = O->maxcmpte;
-    int k;
-    for (i = 0; i < O->maxcmpte; i++){
-        TREEWALK_REDUCE(WINDV_GET_PRIV(tw)->Ngb[pi][i], O->Ngb[i]);
-        TREEWALK_REDUCE(WINDV_GET_PRIV(tw)->V2sum[pi][i], O->V2sum[i]);
-        for(k = 0; k < 3; k ++) {
-            TREEWALK_REDUCE(WINDV_GET_PRIV(tw)->V1sum[pi][i][k], O->V1sum[i][k]);
-        }
-    }
-//     message(1, "Reduce ID=%ld, NGB_first=%d NGB_last=%d maxcmpte = %d, left = %g, right = %g\n",
-//             P[place].ID, O->Ngb[0],O->Ngb[O->maxcmpte-1],WINDP(place, Windd).maxcmpte,WINDP(place, Windd).Left,WINDP(place, Windd).Right);
-}
-
-static void
-wind_vdisp_postprocess(const int i, TreeWalk * tw)
-{
-    const int pi = P[i].PI;
-    const int maxcmpt = WINDV_GET_PRIV(tw)->maxcmpte[pi];
-    int j;
-    double evaldmradius[NWINDHSML];
-    for(j = 0; j < maxcmpt; j++){
-        evaldmradius[j] = vdispeffdmradius(i,j,tw);
-    }
-    int close = 0;
-    WINDV_GET_PRIV(tw)->DMRadius[pi] = ngb_narrow_down(&WINDV_GET_PRIV(tw)->Right[pi], &WINDV_GET_PRIV(tw)->Left[pi], evaldmradius, WINDV_GET_PRIV(tw)->Ngb[pi], maxcmpt, NUMDMNGB, &close, tw->tree->BoxSize);
-    double numngb = WINDV_GET_PRIV(tw)->Ngb[pi][close];
-
-    int tid = omp_get_thread_num();
-    /*  If we have 40 neighbours, or if DMRadius is narrow, set vdisp. Otherwise add to redo queue */
-    if((numngb < (NUMDMNGB - MAXDMDEVIATION) || numngb > (NUMDMNGB + MAXDMDEVIATION)) &&
-        (WINDV_GET_PRIV(tw)->Right[pi] - WINDV_GET_PRIV(tw)->Left[pi] > 1e-2)) {
-        /* More work needed: add this particle to the redo queue*/
-        tw->NPRedo[tid][tw->NPLeft[tid]] = i;
-        tw->NPLeft[tid] ++;
-    }
-    else{
-        double vdisp = WINDV_GET_PRIV(tw)->V2sum[pi][close] / numngb;
-        int d;
-        for(d = 0; d < 3; d++){
-            vdisp -= pow(WINDV_GET_PRIV(tw)->V1sum[pi][close][d] / numngb,2);
-        }
-        if(vdisp > 0) {
-            if(P[i].Type == 0)
-                SPHP(i).VDisp = sqrt(vdisp / 3);
-        }
-    }
-
-    if(tw->maxnumngb[tid] < numngb)
-        tw->maxnumngb[tid] = numngb;
-    if(tw->minnumngb[tid] > numngb)
-        tw->minnumngb[tid] = numngb;
-}
-
-static int
-winds_veldisp_haswork(int n, TreeWalk * tw)
-{
-    /* Don't want a density for swallowed particles*/
-    if(P[n].Swallowed || P[n].IsGarbage)
-        return 0;
-    /* Only want gas*/
-    if(P[n].Type != 0)
-        return 0;
-    /* We only want VDisp for gas particles that may be star-forming over the next PM timestep.
-     * Use DtHsml.*/
-    double densfac = (P[n].Hsml + P[n].DtHsml * WINDV_GET_PRIV(tw)->ddrift)/P[n].Hsml;
-    if(densfac > 1)
-        densfac = 1;
-    if(SPHP(n).Density/(densfac * densfac * densfac) < 0.1 * sfr_density_threshold(WINDV_GET_PRIV(tw)->Time))
-        return 0;
-    /* Update veldisp only on a gravitationally active timestep,
-     * since this is the frequency at which the gravitational acceleration,
-     * which is the only thing DM contributes to, could change. This is probably
-     * overly conservative, because most of the acceleration will be from other stars. */
-//     if(!is_timebin_active(P[n].TimeBinGravity, P[n].Ti_drift))
-//         return 0;
-    return 1;
-}
-
-/* Find the 1D DM velocity dispersion of all gas particles by running a density loop.
- * Stores it in VDisp in the slots structure.*/
-void
-winds_find_vel_disp(const ActiveParticles * act, const double Time, const double hubble, Cosmology * CP, DriftKickTimes * times, DomainDecomp * ddecomp)
-{
-    TreeWalk tw[1] = {0};
-    struct WindVDispPriv priv[1] = {0};
-    ForceTree tree[1] = {0};
-    /* Types used: gas*/
-    tw->ev_label = "WIND_VDISP";
-    tw->fill = (TreeWalkFillQueryFunction) wind_vdisp_copy;
-    tw->reduce = (TreeWalkReduceResultFunction) wind_vdisp_reduce;
-    tw->query_type_elsize = sizeof(TreeWalkQueryWindVDisp);
-    tw->result_type_elsize = sizeof(TreeWalkResultWindVDisp);
-    tw->tree = tree;
-
-    /* sum the total weight of surrounding gas */
-    tw->ngbiter_type_elsize = sizeof(TreeWalkNgbIterWind);
-    tw->ngbiter = (TreeWalkNgbIterFunction) wind_vdisp_ngbiter;
-
-    tw->haswork = winds_veldisp_haswork;
-    tw->visit = (TreeWalkVisitFunction) treewalk_visit_nolist_ngbiter;
-    tw->postprocess = (TreeWalkProcessFunction) wind_vdisp_postprocess;
-
-    priv[0].Time = Time;
-    priv[0].hubble = hubble;
-    priv[0].ddrift = get_exact_drift_factor(CP, times->Ti_Current, times->PM_length);
-    tw->priv = priv;
-
-    /* Build the queue to check that we have something to do before we rebuild the tree.*/
-    treewalk_build_queue(tw, act->ActiveParticle, act->NumActiveParticle, 0);
-
-    int * ActiveVDisp = tw->WorkSet;
-    int64_t NumVDisp = tw->WorkSetSize;
-    int64_t totvdisp, totbh;
-    /* Check for black holes*/
-    MPI_Allreduce(&SlotsManager->info[5].size, &totbh, 1, MPI_INT64, MPI_SUM, MPI_COMM_WORLD);
-    /* If this queue is empty, nothing to do for winds.*/
-    MPI_Allreduce(&NumVDisp, &totvdisp, 1, MPI_INT64, MPI_SUM, MPI_COMM_WORLD);
-
-    if(totvdisp > 0 || totbh > 0) {
-        force_tree_rebuild_mask(tree, ddecomp, DMMASK, NULL);
-        tw->haswork = NULL;
-        init_kick_factor_data(&priv->kf, times, CP);
-    }
-
-    /* Compute the black hole velocity dispersions if needed*/
-    if(totbh)
-        blackhole_veldisp(act, CP, tree, &priv->kf);
-
-    if(totvdisp == 0) {
-        force_tree_free(tree);
-        myfree(ActiveVDisp);
-        return;
-    }
-
-    priv->Left = (MyFloat *) mymalloc("VDISP->Left", SlotsManager->info[0].size * sizeof(MyFloat));
-    priv->Right = (MyFloat *) mymalloc("VDISP->Right", SlotsManager->info[0].size * sizeof(MyFloat));
-    priv->DMRadius = (MyFloat *) mymalloc("VDISP->DMRadius", SlotsManager->info[0].size * sizeof(MyFloat));
-    priv->Ngb = (MyFloat (*) [NWINDHSML]) mymalloc("VDISP->NumNgb", SlotsManager->info[0].size * sizeof(priv->Ngb[0]));
-    priv->V1sum = (MyFloat (*) [NWINDHSML][3]) mymalloc("VDISP->V1Sum", SlotsManager->info[0].size * sizeof(priv->V1sum[0]));
-    priv->V2sum = (MyFloat (*) [NWINDHSML]) mymalloc("VDISP->V2Sum", SlotsManager->info[0].size * sizeof(priv->V2sum[0]));
-    priv->maxcmpte = (int *) mymalloc("maxcmpte", SlotsManager->info[0].size * sizeof(int));
-    report_memory_usage("WIND_VDISP");
-
-    int i;
-    /*Initialise the WINDP array*/
-    #pragma omp parallel for
-    for (i = 0; i < act->NumActiveParticle; i++) {
-        const int n = act->ActiveParticle ? act->ActiveParticle[i] : i;
-        if(P[n].Type == 0) {
-            const int pi = P[n].PI;
-            priv->DMRadius[pi] = P[n].Hsml;
-            priv->Left[pi] = 0;
-            priv->Right[pi] = tree->BoxSize;
-            priv->maxcmpte[pi] = NUMDMNGB;
-        }
-    }
-    /* Find densities*/
-    treewalk_do_hsml_loop(tw, ActiveVDisp, NumVDisp, 1);
-    /* Free memory*/
-    myfree(priv->maxcmpte);
-    myfree(priv->V2sum);
-    myfree(priv->V1sum);
-    myfree(priv->Ngb);
-    myfree(priv->DMRadius);
-    myfree(priv->Right);
-    myfree(priv->Left);
-
-    force_tree_free(tree);
-    myfree(ActiveVDisp);
-    walltime_measure("/Cooling/VDisp");
-
 }
