+++ libgadget/winds.h
@@ -3,7 +3,6 @@
 
 #include "forcetree.h"
 #include "utils/paramset.h"
-#include "timestep.h"
 
 /*
  * Enumeration of the supported wind models.
@@ -48,11 +47,5 @@ void winds_decoupled_hydro(int i, double atime);
 
 /* Returns 1 if the winds ever decouple, 0 otherwise*/
 int winds_ever_decouple(void);
-
-/* Find the 1D DM velocity dispersion of all nearly star-forming gas and black hole particles.
- * Gas is done by running a density loop for find Vdisp of nearest 40 DM particles.
- * BH is done by finding VDisp of all DM particles inside the SPH kernel.
- * Stores it in VDisp in the slots structure.*/
-void winds_find_vel_disp(const ActiveParticles * act, const double Time, const double hubble, Cosmology * CP, DriftKickTimes * times, DomainDecomp * ddecomp);
 
 #endif
